tests/requirements.txt
factory_boy==2.12.0
django==2.2.6
psycopg2==2.7.4
psycopg2-binary==2.7.5
